requirements.txt
flask~=3.0.3
pytest~=8.3.1
loguru~=0.7.2
python-dotenv~=1.0.1
pytelegrambotapi~=4.12.0

# Dev dependencies
mypy~=0.952
pylint~=2.13.5
isort~=5.11.1
black~=23.4.0